remove old code referencing SK
