+++ python/packages/ai/pyproject.toml
@@ -16,6 +16,7 @@ tiktoken = "^0.4.0"
 aiohttp = "^3.8.5"
 jsonschema = "^4.21.1"
 types-pyyaml = "^6.0.12.12"
+pyyaml = "^6.0.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.0"
