+++ python/packages/ai/teams/ai/prompts/__init__.py
@@ -3,8 +3,25 @@ Copyright (c) Microsoft Corporation. All rights reserved.
 Licensed under the MIT License.
 """
 
-from .completion_config import CompletionConfig
+from .assistant_message import AssistantMessage
+from .conversation_history import ConversationHistory
+from .data_source_section import DataSourceSection
+from .function_call import FunctionCall
+from .function_call_message import FunctionCallMessage
+from .function_response_message import FunctionResponseMessage
+from .group_section import GroupSection
+from .layout_engine import LayoutEngine
+from .message import ImageContentPart, ImageUrl, Message, TextContentPart
+from .prompt import Prompt
+from .prompt_functions import PromptFunction, PromptFunctions
 from .prompt_manager import PromptManager
-from .prompt_section import RenderedPromptSection
+from .prompt_manager_options import PromptManagerOptions
+from .prompt_section import PromptSection
+from .prompt_section_base import PromptSectionBase
 from .prompt_template import PromptTemplate
-from .prompt_template_config import PromptTemplateConfig
+from .rendered_prompt_section import RenderedPromptSection
+from .system_message import SystemMessage
+from .template_section import TemplateSection
+from .text_section import TextSection
+from .user_input_message import UserInputMessage
+from .user_message import UserMessage
