+++ python/packages/ai/teams/ai/prompts/assistant_message.py
