+++ python/packages/ai/teams/ai/prompts/function_call_message.py
