+++ python/packages/ai/teams/ai/prompts/function_response_message.py
