+++ python/packages/ai/teams/ai/prompts/group_section.py
