+++ python/packages/ai/teams/ai/prompts/prompt_section_layout.py
