+++ python/packages/ai/teams/ai/prompts/rendered_prompt_section.py
