+++ python/packages/ai/teams/ai/prompts/system_message.py
