+++ python/packages/ai/teams/ai/prompts/template_section.py
