+++ python/packages/ai/teams/ai/prompts/user_input_message.py
