+++ python/packages/ai/teams/ai/prompts/user_message.py
