--- python/packages/ai/teams/ai/promptsv2/__init__.py
@@ -1,27 +1,0 @@
-"""
-Copyright (c) Microsoft Corporation. All rights reserved.
-Licensed under the MIT License.
-"""
-
-from .assistant_message import AssistantMessage
-from .conversation_history import ConversationHistory
-from .data_source_section import DataSourceSection
-from .function_call import FunctionCall
-from .function_call_message import FunctionCallMessage
-from .function_response_message import FunctionResponseMessage
-from .group_section import GroupSection
-from .layout_engine import LayoutEngine
-from .message import ImageContentPart, ImageUrl, Message, TextContentPart
-from .prompt import Prompt
-from .prompt_functions import PromptFunction, PromptFunctions
-from .prompt_manager import PromptManager
-from .prompt_manager_options import PromptManagerOptions
-from .prompt_section import PromptSection
-from .prompt_section_base import PromptSectionBase
-from .prompt_template import PromptTemplate
-from .rendered_prompt_section import RenderedPromptSection
-from .system_message import SystemMessage
-from .template_section import TemplateSection
-from .text_section import TextSection
-from .user_input_message import UserInputMessage
-from .user_message import UserMessage
