--- python/packages/ai/teams/ai/promptsv2/prompt_template.py
@@ -1,27 +1,0 @@
-"""
-Copyright (c) Microsoft Corporation. All rights reserved.
-Licensed under the MIT License.
-"""
-
-from dataclasses import dataclass
-
-from .prompt_section import PromptSection
-from .prompt_template_config import PromptTemplateConfig
-
-
-@dataclass
-class PromptTemplate:
-    """
-    Prompt template cached by the prompt manager.
-
-    Attributes:
-        name (str): Name of the prompt template.
-
-        prompt (PromptSection): Text of the prompt template.
-
-        config (PromptTemplateConfig): Configuration settings for the prompt template.
-    """
-
-    name: str
-    prompt: PromptSection
-    config: PromptTemplateConfig
