+++ python/packages/ai/tests/ai/prompts/test_conversation_history.py
@@ -8,7 +8,7 @@ from unittest.mock import MagicMock
 
 from botbuilder.core import TurnContext
 
-from teams.ai.promptsv2 import ConversationHistory, Message, PromptFunctions
+from teams.ai.prompts import ConversationHistory, Message, PromptFunctions
 from teams.ai.tokenizers import GPTTokenizer
 from teams.state import TurnState
