+++ python/packages/ai/tests/ai/prompts/test_function_response_message.py
@@ -8,7 +8,7 @@ from unittest.mock import MagicMock
 
 from botbuilder.core import TurnContext
 
-from teams.ai.promptsv2 import FunctionResponseMessage, PromptFunctions
+from teams.ai.prompts import FunctionResponseMessage, PromptFunctions
 from teams.ai.tokenizers import GPTTokenizer
 from teams.state import Memory
