+++ python/packages/ai/tests/ai/prompts/test_layout_engine.py
@@ -8,7 +8,7 @@ from unittest import IsolatedAsyncioTestCase
 
 from botbuilder.core import TurnContext
 
-from teams.ai.promptsv2 import LayoutEngine, PromptFunctions, TextSection
+from teams.ai.prompts import LayoutEngine, PromptFunctions, TextSection
 from teams.ai.tokenizers import GPTTokenizer
 from teams.state import TurnState
