+++ python/packages/ai/tests/ai/prompts/test_prompt_section_base.py
@@ -9,7 +9,7 @@ from unittest.mock import AsyncMock
 
 from botbuilder.core import TurnContext
 
-from teams.ai.promptsv2 import (
+from teams.ai.prompts import (
     FunctionCall,
     ImageContentPart,
     Message,
