+++ python/packages/ai/tests/ai/validators/test_json_response_validator.py
@@ -9,7 +9,7 @@ from unittest import IsolatedAsyncioTestCase
 from botbuilder.core import TurnContext
 
 from teams.ai.models import PromptResponse
-from teams.ai.promptsv2 import Message
+from teams.ai.prompts import Message
 from teams.ai.tokenizers import GPTTokenizer
 from teams.ai.validators import JSONResponseValidator
 from teams.state import TurnState
