--- python/packages/ai/tests/prompts/test_assets/happy_path/skprompt.txt
@@ -1,1 +1,0 @@
-test prompt
