--- python/packages/ai/tests/prompts/test_assets/missing_config/skprompt.txt
@@ -1,1 +1,0 @@
-test prompt
