--- python/packages/ai/tests/prompts/test_assets/missing_prompt/config.json
@@ -1,12 +1,0 @@
-{
-    "schema": 1,
-    "description": "test config",
-    "type": "completion",
-    "completion": {
-      "max_tokens": 2000,
-      "temperature": 0.2,
-      "top_p": 0.0,
-      "presence_penalty": 0.0,
-      "frequency_penalty": 0.0
-    }
-}
