+++ python/packages/ai/tests/utils/adapter.py
@@ -4,7 +4,6 @@ Licensed under the MIT License.
 """
 
 from typing import List
-from unittest import IsolatedAsyncioTestCase
 
 from botbuilder.core import BotAdapter
 from botbuilder.schema import Activity, ResourceResponse
@@ -12,25 +11,25 @@ from botbuilder.schema import Activity, ResourceResponse
 from .activity import ACTIVITY
 
 
-class SimpleAdapter(BotAdapter, IsolatedAsyncioTestCase):
+class SimpleAdapter(BotAdapter):
     async def send_activities(self, context, activities) -> List[ResourceResponse]:
         responses = []
 
-        self.assertIsNotNone(context)
-        self.assertIsNotNone(activities)
-        self.assertIsInstance(activities, list)
+        assert context is not None
+        assert activities is not None
+        assert isinstance(activities, list)
 
         for _, activity in enumerate(activities):
-            self.assertIsInstance(activity, Activity)
+            assert isinstance(activity, Activity)
             responses.append(ResourceResponse(id="5678"))
+
         return responses
 
     async def update_activity(self, context, activity):
-        self.assertIsNotNone(context)
-        self.assertIsNotNone(activity)
+        assert context is not None
+        assert activity is not None
         return ResourceResponse(id=activity.id)
 
     async def delete_activity(self, context, reference):
-        self.assertIsNotNone(context)
-        self.assertIsNotNone(reference)
-        self.assertEqual(reference.activity_id, ACTIVITY.id)
+        assert context is not None
+        assert reference is not None
